Browser-based Minim companion...
Work-in-progress but well underway for basic functionalities.
I had a full history but created a new sub-folder repository by mistake :P
Web is much more accessible and agile for small proejcs like this.
Plus Eclipse RPC is always, withou failure, a pain to upgrade.
This might be a premature transition but hopefully the quality of life already jusstifies it.