Changing challenge scaling again...
This time much more confidently. The power-of-5 exponential scale has now emerged from 2 different simulations years-apart (tuktk-encounters and this project's spread-sheet). It's also coincidentally congruent with Minim.
There is some brute-forcing to make the actual linear-challenge fit the pure exponential-curve but it's only a factor of 2 at tier-1 (which is by-far the least-common scenario as tier-1 units are rare and not very fun mechanically) and a strict 20% on higher tiers, which would only become a significant issue for game-play for parties of 5 or more, which is the very upper-bound of typical parties. Even then it's likely off-set by player-advantages such as group-strategy; will-power; items... not to mention it's as likely to benefit the players as other-wise.